# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VERA GATE Payment Gateway API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  operator: 'operator',
  merchant: 'merchant',
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
  /** @nullable */
  merchantId?: number | null;
  createdAt: string;
}

export interface AuthResult {
  user: User;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  admin: 'admin',
  operator: 'operator',
  merchant: 'merchant',
} as const;

export interface UserInput {
  username: string;
  password: string;
  role: UserInputRole;
  /** @nullable */
  merchantId?: number | null;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  admin: 'admin',
  operator: 'operator',
  merchant: 'merchant',
} as const;

export interface UserUpdate {
  username?: string;
  /** @nullable */
  password?: string | null;
  role?: UserUpdateRole;
  /** @nullable */
  merchantId?: number | null;
}

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  MENUNGGU: 'MENUNGGU',
  SUKSES: 'SUKSES',
  GAGAL: 'GAGAL',
  KEDALUWARSA: 'KEDALUWARSA',
} as const;

export interface Transaction {
  id: number;
  ref: string;
  amount: number;
  status: TransactionStatus;
  method: string;
  /** @nullable */
  qrCode?: string | null;
  /** @nullable */
  customerId?: string | null;
  /** @nullable */
  merchantId?: number | null;
  /** @nullable */
  merchantName?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface TransactionList {
  data: Transaction[];
  total: number;
  page: number;
  limit: number;
}

export interface TransactionInput {
  /** @minimum 1000 */
  amount: number;
  customerId: string;
  /** @nullable */
  merchantId?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface TransactionWithQr {
  transaction: Transaction;
  qrCode: string;
  ref: string;
}

export interface CooldownError {
  error: string;
  cooldownMinutes: number;
  remainingMinutes?: number;
  existingRef: string;
  existingCreatedAt: string;
}

export type StatusUpdateStatus = typeof StatusUpdateStatus[keyof typeof StatusUpdateStatus];


export const StatusUpdateStatus = {
  MENUNGGU: 'MENUNGGU',
  SUKSES: 'SUKSES',
  GAGAL: 'GAGAL',
  KEDALUWARSA: 'KEDALUWARSA',
} as const;

export interface StatusUpdate {
  status: StatusUpdateStatus;
  /** @nullable */
  notes?: string | null;
}

export interface CallbackInput {
  /** @nullable */
  order?: string | null;
  /** @nullable */
  orderId?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  transaction_id?: string | null;
  /** @nullable */
  uid?: string | null;
}

export interface StatusCount {
  status: string;
  count: number;
  amount: number;
}

export interface MerchantStat {
  merchantId: number;
  merchantName: string;
  count: number;
  amount: number;
}

export interface DashboardSummary {
  totalTransactions: number;
  totalAmount: number;
  todayCount: number;
  todayAmount: number;
  byStatus: StatusCount[];
  topMerchants?: MerchantStat[];
}

export interface Merchant {
  id: number;
  name: string;
  code: string;
  /** @nullable */
  callbackUrl?: string | null;
  isActive?: boolean;
  createdAt: string;
}

export interface MerchantInput {
  name: string;
  code: string;
  /** @nullable */
  callbackUrl?: string | null;
  isActive?: boolean;
}

export interface MerchantUpdate {
  name?: string;
  code?: string;
  /** @nullable */
  callbackUrl?: string | null;
  isActive?: boolean;
}

export type SettingsFlypayMode = typeof SettingsFlypayMode[keyof typeof SettingsFlypayMode];


export const SettingsFlypayMode = {
  sandbox: 'sandbox',
  live: 'live',
} as const;

export interface Settings {
  id: number;
  flypayAppId: string;
  flypayMode: SettingsFlypayMode;
  callbackBaseUrl: string;
  cooldownMinutes?: number;
  updatedAt: string;
}

export type SettingsUpdateFlypayMode = typeof SettingsUpdateFlypayMode[keyof typeof SettingsUpdateFlypayMode];


export const SettingsUpdateFlypayMode = {
  sandbox: 'sandbox',
  live: 'live',
} as const;

export interface SettingsUpdate {
  flypayAppId?: string;
  /** @nullable */
  flypaySecret?: string | null;
  flypayMode?: SettingsUpdateFlypayMode;
  callbackBaseUrl?: string;
  cooldownMinutes?: number;
}

export type ListTransactionsParams = {
status?: ListTransactionsStatus;
/**
 * @nullable
 */
merchantId?: number | null;
method?: string;
startDate?: string;
endDate?: string;
search?: string;
page?: number;
limit?: number;
};

export type ListTransactionsStatus = typeof ListTransactionsStatus[keyof typeof ListTransactionsStatus];


export const ListTransactionsStatus = {
  MENUNGGU: 'MENUNGGU',
  SUKSES: 'SUKSES',
  GAGAL: 'GAGAL',
  KEDALUWARSA: 'KEDALUWARSA',
} as const;

export type ListSuksesTransactionsParams = {
/**
 * @nullable
 */
merchantId?: number | null;
startDate?: string;
endDate?: string;
page?: number;
limit?: number;
};

export type GetDashboardSummaryParams = {
period?: GetDashboardSummaryPeriod;
};

export type GetDashboardSummaryPeriod = typeof GetDashboardSummaryPeriod[keyof typeof GetDashboardSummaryPeriod];


export const GetDashboardSummaryPeriod = {
  today: 'today',
  week: 'week',
  month: 'month',
  all: 'all',
} as const;

export type GetDashboardRecentParams = {
limit?: number;
};