# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VERA GATE Payment Gateway API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /merchants:
    get:
      operationId: listMerchants
      tags: [merchants]
      summary: List all merchants
      responses:
        "200":
          description: List of merchants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Merchant"
    post:
      operationId: createMerchant
      tags: [merchants]
      summary: Create a new merchant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchantInput"
      responses:
        "201":
          description: Created merchant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
  /merchants/{id}:
    patch:
      operationId: updateMerchant
      tags: [merchants]
      summary: Update a merchant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchantUpdate"
      responses:
        "200":
          description: Updated merchant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
    delete:
      operationId: deleteMerchant
      tags: [merchants]
      summary: Delete a merchant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, operator, merchant]
        merchantId:
          type: ["integer", "null"]
        createdAt:
          type: string
...
    Merchant:
      type: object
      required: [id, name, code, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        callbackUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    MerchantInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        callbackUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
    MerchantUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        callbackUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
...